Refactor: test ssg on share page
